Move shared-web models to main application. It appears the DateConcern is not used in the cosmetics-web application, so it was removed
