--- shared-web/app/models/shared/web/application_record.rb
@@ -1,7 +1,0 @@
-module Shared
-  module Web
-    class ApplicationRecord < ActiveRecord::Base
-      self.abstract_class = true
-    end
-  end
-end
